Bug fix -- multiple values for argument.
The source of this bug is that if defaults were present for an argument, and the
argument as provided on the command line in the verbose syntax (with --), then
the value would be supplied with both the default value, and as a keyword argument
to the function. The solution was to just go through arguments, and if there ist
a corresponding argument in keyword arguments, to pop it off and put it in.