Windows support
In setup.py, switch scripts to entry_points so that a shovel-script.py and
shovel.exe will be generated.

In shovel/runner.py, explicitly set the `prog` argument to the argparser
so that the usage shows shovel instead of shovel-script.py.

In the tests, normalize the paths to fix forward slash vs backslash
conflicts.

NOTE: There is still an error in one of the tests. That test doesn't seem
special, so I'm not sure why that one in particular is causing trouble. I
will look into it further.